requirements.txt
langchain>=0.3.26          
langchain-community>=0.0.30 
langchain-groq>=0.3.5         
faiss-cpu>=1.7.4            
sentence-transformers>=2.2.2  
PyPDF2>=3.0.1  
python-dotenv>=1.0.1  
google-generativeai>=0.5.0  
playwright>=1.44.0         
asyncio  
feedparser  
bs4  
# cloudscraper>=3.0.0    
newspaper4k  
beautifulsoup4>=4.12.3  
requests>=2.31.0  
scikit-learn>=1.3.0  
lxml>=5.2.0  
tqdm>=4.66.4
openai>=1.30.1
langchain>=0.1.20
python-dotenv>=1.0.1
reportlab>=4.0.5
python-telegram-bot
schedule